pyproject.toml
[project]
name = "simple-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.122.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
